# Make ShortestPath faster