I noticed that the performance of `ShortestPath` can be substantially improved. Instead of finding all shortest paths from the initial vertex, then picking the shortest path between _v_ and _w_ we want, we could terminate the search as soon as we found the path to the desired vertex.

I've opened a pull request, because I already have a patch.

1. No API changes whatsoever;
2. This performance enhancement is based on the original E. W. Dijkstra's paper (it's what should be found in a CS textbook).

old/new comparison:

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkShortestPath250-4      55089         31078         -43.59%
BenchmarkShortestPath500-4      121956        51563         -57.72%
BenchmarkShortestPath1000-4     206507        79512         -61.50%
```