@amwolff 2020-08-27
I was going through my old contributions and realized I made a mistake here in the comment about `ShortestPath` (it should be kept original). I've fixed this in case this may get merged. :)